refactor: rename a folder slices to reducers
