+++ react-app/src/store/reducers/charactersSlice.ts
