+++ react-app/src/store/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { charactersApi } from '../api/charactersApi';
-import charactersReducer from './slices/charactersSlice';
-import carsSlice from './slices/carsSlice';
+import charactersReducer from './reducers/charactersSlice';
+import carsSlice from './reducers/carsSlice';
 
 export const store = configureStore({
   reducer: {
